requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.5
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
idna==3.4
ipdb==0.13.11
ipython==8.8.0
jedi==0.18.2
jwcrypto==1.4.2
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
numpy==1.24.1
oauthlib==3.2.2
pandas==1.5.3
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.1
urllib3==1.26.14
wcwidth==0.2.6
wrapt==1.14.1
